+++ Client/Pages/DetailView/EditPurchaseOrder.razor.cs
@@ -125,7 +125,7 @@ namespace Vanigam.CRM.Client.Pages.DetailView
             }
         }
 
-        private async Task OnDiscountPercentageChanged(double discountPercent)
+        private async Task OnDiscountPercentageChanged(decimal discountPercent)
         {
             if (CurrentObject != null)
             {
